PR feedback + general cleanup
