+++ packages/web/src/pages/sign-on/components/mobile/SignOnContainerMobile.tsx
@@ -1,0 +1,11 @@
+import { PropsWithChildren } from 'react'
+
+import { Flex } from '@audius/harmony'
+
+export const SignOnContainerMobile = ({ children }: PropsWithChildren<{}>) => {
+  return (
+    <Flex direction='column' h='100%' css={{ overflow: 'hidden' }}>
+      {children}
+    </Flex>
+  )
+}
